requirements.txt
# Core dependencies with compatible versions
gradio>=3.50.0
Pillow>=8.0.0
PyYAML>=5.4.0
ffmpy>=0.2.2
tqdm>=4.60.0

# NumPy - use compatible version for all packages
numpy>=1.21.0,<2.0.0

# PyTorch ecosystem - use compatible versions (avoid 2.6 for TTS compatibility)
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2

# Scientific computing
scipy>=1.7.0,<2.0.0

# Computer vision
opencv-python>=4.5.0,<5.0.0

# Audio processing
soundfile>=0.10.0,<1.0.0
librosa>=0.8.0,<1.0.0
pyttsx3>=2.90

# Visualization
matplotlib>=3.3.0,<4.0.0

# Voice cloning - install last to avoid conflicts
TTS>=0.22.0,<0.25.0
transformers>=4.30.0,<4.40.0
accelerate>=0.20.0,<1.0.0 